ARROW-14244: [C++] Reduce scalar_temporal.cc compilation time
* Split unary and binary kernels into two distinct translation units
* Fix temporal codegen infrastructure so as not to generate code for kernels
  for unused argument types (e.g. avoid generating code for time inputs
  if a kernel is only defined for date types)
