+++ cpp/src/arrow/compute/kernels/temporal_internal.h
@@ -20,6 +20,7 @@
 #include <chrono>
 #include <cstdint>
 
+#include "arrow/compute/api_scalar.h"
 #include "arrow/vendored/datetime.h"
 
 namespace arrow {
@@ -67,6 +68,15 @@ static inline const std::string& GetInputTimezone(const ArrayData& array) {
   return checked_cast<const TimestampType&>(*array.type).timezone();
 }
 
+inline Status ValidateDayOfWeekOptions(const DayOfWeekOptions& options) {
+  if (options.week_start < 1 || 7 < options.week_start) {
+    return Status::Invalid(
+        "week_start must follow ISO convention (Monday=1, Sunday=7). Got week_start=",
+        options.week_start);
+  }
+  return Status::OK();
+}
+
 struct NonZonedLocalizer {
   using days_t = sys_days;
 
@@ -92,6 +102,48 @@ struct ZonedLocalizer {
 
   local_days ConvertDays(sys_days d) const { return local_days(year_month_day(d)); }
 };
+
+//
+// Which types to generate a kernel for
+//
+struct WithDates {};
+struct WithTimes {};
+struct WithTimestamps {};
+
+// This helper allows generating temporal kernels for selected type categories
+// without any spurious code generation for other categories (e.g. avoid
+// generating code for date kernels for a times-only function).
+template <typename Factory>
+void AddTemporalKernels(Factory* fac) {}
+
+template <typename Factory, typename... WithOthers>
+void AddTemporalKernels(Factory* fac, WithDates, WithOthers... others) {
+  fac->template AddKernel<days, Date32Type>(date32());
+  fac->template AddKernel<std::chrono::milliseconds, Date64Type>(date64());
+  AddTemporalKernels(fac, std::forward<WithOthers>(others)...);
+}
+
+template <typename Factory, typename... WithOthers>
+void AddTemporalKernels(Factory* fac, WithTimes, WithOthers... others) {
+  fac->template AddKernel<std::chrono::seconds, Time32Type>(time32(TimeUnit::SECOND));
+  fac->template AddKernel<std::chrono::milliseconds, Time32Type>(time32(TimeUnit::MILLI));
+  fac->template AddKernel<std::chrono::microseconds, Time64Type>(time64(TimeUnit::MICRO));
+  fac->template AddKernel<std::chrono::nanoseconds, Time64Type>(time64(TimeUnit::NANO));
+  AddTemporalKernels(fac, std::forward<WithOthers>(others)...);
+}
+
+template <typename Factory, typename... WithOthers>
+void AddTemporalKernels(Factory* fac, WithTimestamps, WithOthers... others) {
+  fac->template AddKernel<std::chrono::seconds, TimestampType>(
+      match::TimestampTypeUnit(TimeUnit::SECOND));
+  fac->template AddKernel<std::chrono::milliseconds, TimestampType>(
+      match::TimestampTypeUnit(TimeUnit::MILLI));
+  fac->template AddKernel<std::chrono::microseconds, TimestampType>(
+      match::TimestampTypeUnit(TimeUnit::MICRO));
+  fac->template AddKernel<std::chrono::nanoseconds, TimestampType>(
+      match::TimestampTypeUnit(TimeUnit::NANO));
+  AddTemporalKernels(fac, std::forward<WithOthers>(others)...);
+}
 
 //
 // Executor class for temporal component extractors, i.e. scalar kernels
