+++ cpp/src/arrow/compute/registry.cc
@@ -161,7 +161,8 @@ static std::unique_ptr<FunctionRegistry> CreateBuiltInRegistry() {
   RegisterScalarStringAscii(registry.get());
   RegisterScalarValidity(registry.get());
   RegisterScalarIfElse(registry.get());
-  RegisterScalarTemporal(registry.get());
+  RegisterScalarTemporalBinary(registry.get());
+  RegisterScalarTemporalUnary(registry.get());
 
   RegisterScalarOptions(registry.get());
